feat(auth): Use temporary credentials for CI runs
This feature introduces the option for integration testing of aws-ssm in
a pipeline that has the workflow of assuming a role and injecting the
credentials as chart values. With the role-arn annotation, the
expectation is that you run on EC2, with this approach, any pipeline can
create temporary credentials and inject them. Verified and tested on a
kind cluster with a local setup.